# swagger.yaml
swagger: '2.0'
info:
  description: 'Assisted installation'
  version: 1.0.0
  title: AssistedInstall
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: api.openshift.com
basePath: /api/assisted-install
...
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  agentAuth:
    type: apiKey
    in: header
    name: X-Secret-Key
  userAuth:
    type: apiKey
    in: header
    name: Authorization
  urlAuth:
    type: apiKey
    in: query
    name: api_key
security:
  - userAuth: [admin, user]
paths:
...
  /v2/clusters:
    post:
      tags:
        - installer
      description: Creates a new OpenShift cluster definition.
      operationId: v2RegisterCluster
      parameters:
        - in: body
          name: new-cluster-params
          description: The properties describing the new cluster.
          required: true
          schema:
            $ref: '#/definitions/cluster-create-params'
      responses:
        "201":
          description: Success.
          schema:
            $ref: '#/definitions/cluster'
        "400":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized.
          schema:
            $ref: '#/definitions/infra_error'
        "403":
          description: Forbidden.
          schema:
            $ref: '#/definitions/infra_error'
        "405":
          description: Method Not Allowed.
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Error.
          schema:
            $ref: '#/definitions/error'

    get:
      tags:
        - installer
      security:
        - userAuth: [admin, read-only-admin, user]
      description: Retrieves the list of OpenShift clusters.
      operationId: v2ListClusters
      parameters:
        - in: header
          name: get_unregistered_clusters
          description: Whether to return clusters that have been unregistered.
          type: boolean
          required: false
          default: false
        - in: query
          name: openshift_cluster_id
          description: A specific cluster to retrieve.
          type: string
          format: uuid
          required: false
        - in: query
          name: ams_subscription_ids
          description: If non-empty, returned Clusters are filtered to those with matching subscription IDs.
          required: false
          type: array
          items:
            type: string
        - in: query
          name: with_hosts
          description: Include hosts in the returned list.
          type: boolean
          allowEmptyValue: true
          default: false
      responses:
        "200":
          description: Success.
          schema:
            $ref: '#/definitions/cluster-list'
        "401":
          description: Unauthorized.
          schema:
            $ref: '#/definitions/infra_error'
        "403":
          description: Forbidden.
          schema:
            $ref: '#/definitions/infra_error'
        "405":
          description: Method Not Allowed.
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "503":
          description: Unavailable.
          schema:
            $ref: '#/definitions/error'
  /v1/clusters/{cluster_id}:
    get:
      tags:
        - installer
      security:
        - userAuth: [admin, read-only-admin, user]
        - agentAuth: []
      description: Retrieves the details of the OpenShift cluster.
      operationId: GetCluster
      parameters:
        - in: path
          name: cluster_id
          description: The cluster to be retrieved.
          type: string
          format: uuid
          required: true
        - in: header
          name: discovery_agent_version
          description: The software version of the discovery agent that is retrieving the cluster details.
          type: string
          required: false
        - in: header
          name: get_unregistered_clusters
          description: Whether to return clusters that have been unregistered.
          type: boolean
          required: false
          default: false
      responses:
        "200":
          description: Success.
          schema:
            $ref: '#/definitions/cluster'
        "401":
          description: Unauthorized.
          schema:
            $ref: '#/definitions/infra_error'
        "403":
          description: Forbidden.
          schema:
            $ref: '#/definitions/infra_error'
        "404":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "405":
          description: Method Not Allowed.
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "503":
          description: Unavailable.
          schema:
            $ref: '#/definitions/error'

    patch:
      tags:
        - installer
      description: Updates an OpenShift cluster definition.
      operationId: UpdateCluster
      parameters:
        - in: path
          name: cluster_id
          description: The cluster to be updated.
          type: string
          format: uuid
          required: true
        - in: body
          name: cluster-update-params
          description: The properties to update.
          required: true
          schema:
            $ref: '#/definitions/cluster-update-params'
      responses:
        "201":
          description: Success.
          schema:
            $ref: '#/definitions/cluster'
        "400":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized.
          schema:
            $ref: '#/definitions/infra_error'
        "403":
          description: Forbidden.
          schema:
            $ref: '#/definitions/infra_error'
        "404":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "405":
          description: Method Not Allowed.
          schema:
            $ref: '#/definitions/error'
        "409":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Error.
          schema:
            $ref: '#/definitions/error'

    delete:
      tags:
        - installer
      description: Deletes an OpenShift cluster definition.
      operationId: DeregisterCluster
      parameters:
        - in: path
          name: cluster_id
          description: The cluster to be deregistered.
          format: uuid
          type: string
          required: true
      responses:
        "204":
          description: Success.
        "401":
          description: Unauthorized.
          schema:
            $ref: '#/definitions/infra_error'
        "403":
          description: Forbidden.
          schema:
            $ref: '#/definitions/infra_error'
        "404":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "405":
          description: Method Not Allowed.
          schema:
            $ref: '#/definitions/error'
        "409":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Error.
          schema:
            $ref: '#/definitions/error'
  /v2/clusters/{cluster_id}:
    get:
      tags:
        - installer
      security:
        - userAuth: [admin, read-only-admin, user]
        - agentAuth: []
      description: Retrieves the details of the OpenShift cluster.
      operationId: v2GetCluster
      parameters:
        - in: path
          name: cluster_id
          description: The cluster to be retrieved.
          type: string
          format: uuid
          required: true
        - in: header
          name: discovery_agent_version
          description: The software version of the discovery agent that is retrieving the cluster details.
          type: string
          required: false
        - in: header
          name: get_unregistered_clusters
          description: Whether to return clusters that have been unregistered.
          type: boolean
          required: false
          default: false
      responses:
        "200":
          description: Success.
          schema:
            $ref: '#/definitions/cluster'
        "401":
          description: Unauthorized.
          schema:
            $ref: '#/definitions/infra_error'
        "403":
          description: Forbidden.
          schema:
            $ref: '#/definitions/infra_error'
        "404":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "405":
          description: Method Not Allowed.
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "503":
          description: Unavailable.
          schema:
            $ref: '#/definitions/error'

    delete:
      tags:
        - installer
      description: Deletes an OpenShift cluster definition.
      operationId: v2DeregisterCluster
      parameters:
        - in: path
          name: cluster_id
          description: The cluster to be deregistered.
          format: uuid
          type: string
          required: true
      responses:
        "204":
          description: Success.
        "401":
          description: Unauthorized.
          schema:
            $ref: '#/definitions/infra_error'
        "403":
          description: Forbidden.
          schema:
            $ref: '#/definitions/infra_error'
        "404":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "405":
          description: Method Not Allowed.
          schema:
            $ref: '#/definitions/error'
        "409":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Error.
          schema:
            $ref: '#/definitions/error'

    patch:
      tags:
        - installer
      description: Updates an OpenShift cluster definition.
      operationId: V2UpdateCluster
      parameters:
        - in: path
          name: cluster_id
          description: The cluster to be updated.
          type: string
          format: uuid
          required: true
        - in: body
          name: cluster-update-params
          description: The properties to update.
          required: true
          schema:
            $ref: '#/definitions/v2-cluster-update-params'
      responses:
        "201":
          description: Success.
          schema:
            $ref: '#/definitions/cluster'
        "400":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized.
          schema:
            $ref: '#/definitions/infra_error'
        "403":
          description: Forbidden.
          schema:
            $ref: '#/definitions/infra_error'
        "404":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "405":
          description: Method Not Allowed.
          schema:
            $ref: '#/definitions/error'
        "409":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Error.
          schema:
            $ref: '#/definitions/error'
...
  /v1/clusters/default-config:
    get:
      tags:
        - installer
      security:
        - userAuth: [ admin, read-only-admin, user ]
      description: Get the default values for various cluster properties.
      operationId: GetClusterDefaultConfig
      responses:
        "200":
          description: Success.
          schema:
            $ref: '#/definitions/cluster_default_config'
        "401":
          description: Unauthorized.
          schema:
            $ref: '#/definitions/infra_error'
        "403":
          description: Forbidden.
          schema:
            $ref: '#/definitions/infra_error'
        "500":
          description: Error.
          schema:
            $ref: '#/definitions/error'
...
  /v1/clusters/{cluster_id}/uploads/ingress-cert:
    post:
      tags:
        - installer
      security:
        - agentAuth: []
      description: Transfer the ingress certificate for the cluster.
      operationId: UploadClusterIngressCert
      parameters:
        - in: path
          name: cluster_id
          description: The cluster to associate with the ingress certificate.
          type: string
          format: uuid
          required: true
        - in: body
          name: ingress-cert-params
          description: The ingress certificate.
          required: true
          schema:
            $ref: '#/definitions/ingress-cert-params'
        - in: header
          name: discovery_agent_version
          description: The software version of the discovery agent that is uploading the ingress certificate.
          type: string
          required: false
      responses:
        "201":
          description: Success.
        "400":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized.
          schema:
            $ref: '#/definitions/infra_error'
        "403":
          description: Forbidden.
          schema:
            $ref: '#/definitions/infra_error'
        "404":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "405":
          description: Method Not Allowed.
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "503":
          description: Unavailable.
          schema:
            $ref: '#/definitions/error'
...
  /v1/clusters/{cluster_id}/actions/reset:
    post:
      tags:
        - installer
      description: Resets a failed installation.
      operationId: ResetCluster
      parameters:
        - in: path
          name: cluster_id
          description: The cluster whose installation is to be reset.
          type: string
          format: uuid
          required: true
      responses:
        "202":
          description: Success.
          schema:
            $ref: '#/definitions/cluster'
        "401":
          description: Unauthorized.
          schema:
            $ref: '#/definitions/infra_error'
        "403":
          description: Forbidden.
          schema:
            $ref: '#/definitions/infra_error'
        "404":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "405":
          description: Method Not Allowed.
          schema:
            $ref: '#/definitions/error'
        "409":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Error.
          schema:
            $ref: '#/definitions/error'
  /v1/clusters/{cluster_id}/actions/complete_installation:
    post:
      tags:
        - installer
      security:
        - agentAuth: []
      description: Agent API to mark a finalizing installation as complete.
      operationId: CompleteInstallation
      parameters:
        - in: path
          name: cluster_id
          description: The cluster whose installation is being completing.
          type: string
          format: uuid
          required: true
        - in: body
          name: completion-params
          description: The final status of the cluster installation.
          required: true
          schema:
            $ref: '#/definitions/completion-params'
        - in: header
          name: discovery_agent_version
          description: The software version of the discovery agent that is completing the installation.
          type: string
          required: false
      responses:
        "202":
          description: Success.
          schema:
            $ref: '#/definitions/cluster'
        "401":
          description: Unauthorized.
          schema:
            $ref: '#/definitions/infra_error'
        "403":
          description: Forbidden.
          schema:
            $ref: '#/definitions/infra_error'
        "404":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "405":
          description: Method Not Allowed.
          schema:
            $ref: '#/definitions/error'
        "409":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "500":
          description: Error.
          schema:
            $ref: '#/definitions/error'
        "503":
          description: Unavailable.
          schema:
            $ref: '#/definitions/error'
...
  /v2/clusters/default-config:
    get:
      tags:
        - installer
      security:
        - userAuth: [ admin, read-only-admin, user ]
      description: Get the default values for various cluster properties.
      operationId: V2GetClusterDefaultConfig
      responses:
        "200":
          description: Success.
          schema:
            $ref: '#/definitions/cluster_default_config'
        "401":
          description: Unauthorized.
          schema:
            $ref: '#/definitions/infra_error'
        "403":
          description: Forbidden.
          schema:
            $ref: '#/definitions/infra_error'
        "500":
          description: Error.
          schema:
            $ref: '#/definitions/error'
...
definitions:
...
  cluster-list:
    type: array
    items:
      $ref: '#/definitions/cluster'
...
  ingress-cert-params:
    type: string
  completion-params:
    type: object
    required:
      - is_success
    properties:
      is_success:
        type: boolean
      error_info:
        type: string
...
  error:
    type: object
    required:
      - kind
      - id
      - href
      - code
      - reason
    properties:
      kind:
        type: string
        enum: ['Error']
        description: Indicates the type of this object. Will always be 'Error'.
      id:
        type: integer
        format: int32
        description: Numeric identifier of the error.
        minimum: 400
        maximum: 504
      href:
        type: string
        description: Self link.
      code:
        type: string
        description: Globally unique code of the error, composed of the unique identifier of the API and the numeric identifier of the error. For example, if the numeric identifier of the error is 93 and the identifier of the API is assisted_install then the code will be ASSISTED-INSTALL-93.
      reason:
        type: string
        description: Human-readable description of the error.
...
  infra_error:
    type: object
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
        description: Numeric identifier of the error.
        minimum: 401
        maximum: 403
      message:
        type: string
        description: Human-readable description of the error.
...